Updated read files
Added the possibility to omit the file extension. EX.: python basex.py --file bases.txt or python basex.py --file bases
Note: If you provide a .bin file, the code will attempt to read the contents of that file. However, base conversion depends on correctly interpreting the data in the file. If the file contains data that cannot be interpreted as numbers or is not formatted correctly for base conversion, errors may occur during program execution. Be sure to provide properly formatted text files to avoid conversion errors.